Выносим отдельно утилитарный модуль util.js
