+++ src/util.js
@@ -1,0 +1,4 @@
+const render = (container, template, position = `beforeend`) =>
+  container.insertAdjacentHTML(position, template);
+
+export {render};
