+++ src/view/load-button.js
@@ -1,0 +1,4 @@
+const createButtonLoadMoreTemplate = () =>
+  `<button class="load-more" type="button">load more</button>`;
+
+export {createButtonLoadMoreTemplate};
